chore(realtime,docs,cli): finish Option C cleanup + rewrite talk materials
- shared/core: install WebSockets in HttpClient; widen BaseViewModel.uiState
  to public so Kotlin consumers can collectAsState()
- shared/realtime: Vehicle.lastUpdateMillis Long -> Double for JS export
- demo/realtime-mobile: drop duplicated data/domain/repo/VM, bind to the
  library types; thin AppModule over initKoin(realtimeConfig); AGP 9 bypass
- demo/realtime-api: remove dead JS (seed.js, dev-start.js, *.model.js)
- demo/node-api: initializeWithRealtime + /modules/active-trips
- sl-cli: realtime indicator in status + publish hints
- docs: rewrite SLIDES_PRESENTER_NOTES.md to the coherent 23-slide deck,
  add DECK_REBUILD_SPEC.md build manifest, refresh FEATURES/README/PUBLISHING/CLAUDE

+++ shared/realtime/src/commonMain/kotlin/com/umain/transport/realtime/domain/model/Vehicle.kt
@@ -27,5 +27,12 @@ data class Vehicle(
     val direction: Int,
     val longitude: Double,
     val latitude: Double,
-    val lastUpdateMillis: Long,
+    /**
+     * Epoch ms as Double — Kotlin/JS in 2.3 can't @JsExport `Long` without
+     * BigInt support, and BigInt breaks `JSON.stringify` for the same
+     * reason `Line.transportMode` is a String instead of a Kotlin enum.
+     * Double has 53 bits of integer precision, so timestamps stay accurate
+     * past the year 287,000.
+     */
+    val lastUpdateMillis: Double,
 )
